requirements.txt
pyaudio
openai-whisper
torch
numpy
